// projects/netgrif-application-engine/src/lib/task-content/model/data-field-type.ts
/**
 * Contains types of renderable data fields.
 */
export enum DataFieldType {
    BOOLEAN = 'boolean',
    BUTTON = 'button',
    TEXT = 'text',
    ENUMERATION = 'enumeration',
    NUMBER = 'number',
    MULTICHOICE = 'multichoice',
    DATE = 'date',
    DATE_TIME = 'dateTime',
    FILE = 'file',
    FILE_LIST = 'fileList',
    USER = 'user'
}